+++ tests/test_browser_context.py
@@ -18,17 +18,17 @@ class TestBrowserContext(BaseTestCase):
     async def test_default_context(self):
         self.assertEqual(len(self.browser.browserContexts), 1)
         defaultContext = self.browser.browserContexts[0]
-        self.assertFalse(defaultContext.isIncognite())
+        self.assertFalse(defaultContext.isIncognito())
         with self.assertRaises(BrowserError) as cm:
             await defaultContext.close()
         self.assertIn('cannot be closed', cm.exception.args[0])
 
     @unittest.skip('this test not pass in some environment')
     @sync
-    async def test_incognite_context(self):
+    async def test_incognito_context(self):
         self.assertEqual(len(self.browser.browserContexts), 1)
-        context = await self.browser.createIncogniteBrowserContext()
-        self.assertTrue(context.isIncognite())
+        context = await self.browser.createIncognitoBrowserContext()
+        self.assertTrue(context.isIncognito())
         self.assertEqual(len(self.browser.browserContexts), 2)
         self.assertIn(context, self.browser.browserContexts)
         await context.close()
@@ -37,7 +37,7 @@ class TestBrowserContext(BaseTestCase):
     @sync
     async def test_close_all_targets_once(self):
         self.assertEqual(len(await self.browser.pages()), 2)
-        context = await self.browser.createIncogniteBrowserContext()
+        context = await self.browser.createIncognitoBrowserContext()
         await context.newPage()
         self.assertEqual(len(await self.browser.pages()), 3)
         await context.close()
@@ -45,7 +45,7 @@ class TestBrowserContext(BaseTestCase):
 
     @sync
     async def test_window_open_use_parent_tab_context(self):
-        context = await self.browser.createIncogniteBrowserContext()
+        context = await self.browser.createIncognitoBrowserContext()
         page = await context.newPage()
         await page.goto(self.url + 'empty')
         asyncio.ensure_future(
@@ -56,7 +56,7 @@ class TestBrowserContext(BaseTestCase):
 
     @sync
     async def test_fire_target_event(self):
-        context = await self.browser.createIncogniteBrowserContext()
+        context = await self.browser.createIncognitoBrowserContext()
         events = []
         context.on('targetcreated', lambda t: events.append('CREATED: ' + t.url))  # noqa: E501
         context.on('targetchanged', lambda t: events.append('CHANGED: ' + t.url))  # noqa: E501
@@ -73,12 +73,12 @@ class TestBrowserContext(BaseTestCase):
     @unittest.skip('this test not pass in some environment')
     @sync
     async def test_isolate_loacal_storage_and_cookie(self):
-        context1 = await self.browser.createIncogniteBrowserContext()
-        context2 = await self.browser.createIncogniteBrowserContext()
+        context1 = await self.browser.createIncognitoBrowserContext()
+        context2 = await self.browser.createIncognitoBrowserContext()
         self.assertEqual(len(context1.targets()), 0)
         self.assertEqual(len(context2.targets()), 0)
 
-        # create a page in the first incognite context
+        # create a page in the first incognito context
         page1 = await context1.newPage()
         await page1.goto(self.url + 'empty')
         await page1.evaluate('''() => {
@@ -89,7 +89,7 @@ class TestBrowserContext(BaseTestCase):
         self.assertEqual(len(context1.targets()), 1)
         self.assertEqual(len(context2.targets()), 0)
 
-        # create a page in the second incognite context
+        # create a page in the second incognito context
         page2 = await context2.newPage()
         await page2.goto(self.url + 'empty')
         await page2.evaluate('''() => {
@@ -115,7 +115,7 @@ class TestBrowserContext(BaseTestCase):
     @sync
     async def test_accross_session(self):
         self.assertEqual(len(self.browser.browserContexts), 1)
-        context = await self.browser.createIncogniteBrowserContext()
+        context = await self.browser.createIncognitoBrowserContext()
         self.assertEqual(len(self.browser.browserContexts), 2)
         remoteBrowser = await connect(
             browserWSEndpoint=self.browser.wsEndpoint)